Added multithreading

+++ InOneWeekend/hittable.h
@@ -9,7 +9,7 @@ struct hit_record{
     point3 p;
     vec3 normal;
     shared_ptr<material> mat_ptr;
-    double t;
+    float t;
     bool front_face;
 
     inline void set_face_normal(const ray& r, const vec3& outward_normal){
@@ -20,7 +20,7 @@ struct hit_record{
 
 class hittable{
     public:
-        virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const=0;  
+        virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const=0;  
 };
 
 #endif
